coordonnées en double et plus en int

+++ app/src/main/java/com/towerint/tower.java
@@ -1,8 +1,8 @@
 package com.towerint;
 
 abstract class tower {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private int radius;
     private int range;
     private int speedAttack;
@@ -12,6 +12,5 @@ abstract class tower {
     private int popularity_max;
     private int ProbabilityLoosePopularity;
     // TODO Eventuellement variable qui définit quelles sont les cibles de la tour
-
 
 }
